Change code structure for artifact
