+++ src/listener/listener.js
@@ -20,7 +20,7 @@ import requirement from "./requirement.js";
 import operation_def from "./operation_def.js";
 import interface_type from "./interface_type.js";
 import artifact_type from "./ArtifactTypeListener.js";
-import artifact from "./artifact.js";
+import artifact from "./ArtifactListener.js";
 import implementation from "./implementation.js";
 import group_type from "./group_type.js";
 export default {
